renamed spinner component to loading spinner component, moved to core
