+++ client/src/app/core/loading-spinner/loading-spinner.component.html
