+++ client/src/app/core/loading-spinner/loading-spinner.component.ts
@@ -1,16 +1,16 @@
 import { Component } from '@angular/core';
-import { LoadingService } from '../../core/loading-service/loading.service';
+import { LoadingService } from '../loading-service/loading.service';
 import { Observable } from 'rxjs';
 import { AsyncPipe } from '@angular/common';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 
 @Component({
-  selector: 'app-spinner',
-  templateUrl: './spinner.component.html',
-  styleUrls: ['./spinner.component.css'],
+  selector: 'app-loading-spinner',
+  templateUrl: './loading-spinner.component.html',
+  styleUrls: ['./loading-spinner.component.css'],
   imports: [MatProgressSpinnerModule, AsyncPipe]
 })
-export class SpinnerComponent {
+export class LoadingSpinnerComponent {
 
   loading$: Observable<boolean>;
