+++ client/src/app/core/toolbar-sidenav/toolbar-sidenav.component.html
@@ -104,7 +104,7 @@
       </mat-nav-list>
     </mat-sidenav>
     <mat-sidenav-content class="main-content">
-      <app-spinner />
+      <app-loading-spinner />
       <router-outlet />
     </mat-sidenav-content>
   </mat-sidenav-container>
